"Fix" the test for calling remove_file on directories.
I still have no idea what's happening here, whether it's
macOS doing weird things, Python doing weird things, or
other.

man 3 remove seems to say that calling it on a directory
is the same as rmdir, so I'd expect a ENOTEMPTY here if
that's being called, but instead you seem to get EPERM.

On macOS though, I don't see man 2 unlink saying that it
gives you that if you call it on a directory (even an
empty one).