Almost there. Break out symbolic loop tests on their own.
Still broken, but now I am pretty sure the only broken
part is getting the right path to show up in error messages
when doing operations on childs of looped symbolic links.

That and the last remaining failing test in realpath, which
thankfully now fails the same way for both native and memory.