Make all the important realpath tests pass.
I'm still a bit lacking in confidence :/

But this does actually seem to green things, other than
is_link, which is a bit of a sore thumb at the minute
for not checking for loops.

To fix the latter requires introducing lstat, which
can go in imminently.

Man this commit message rambles no sense but it's New
Years and who gives a damn.